notification_service/pyproject.toml
[tool.poetry]
name = "notification-service"
version = "0.1.0"
description = ""
authors = ["Mostafa <mostafa.babaii@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
djangorestframework = "^3.15.2"
pre-commit = "^3.7.1"
psycopg2-binary = "^2.9.9"
django-environ = "^0.11.2"
djangorestframework-simplejwt = "^5.3.1"
celery = "^5.4.0"
redis = "^5.0.7"
pika = "^1.3.2"
gunicorn = "^22.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
